requirements.txt
flask
flask-cors
python-dotenv
requests
langchain
langchain-community
pymupdf